requirements.txt
django
djangorestframework
djangorestframework-simplejwt
celery
redis
psycopg2-binary
